requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
Django==3.1.1
djangorestframework==3.12.1
filelock==3.0.12
flake8==3.8.3
idna==2.10
mccabe==0.6.1
pendulum==2.1.2
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
pytzdata==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.31
